sstrings can now be printed to TerminalStream directly
